# api/spring-aop-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"OwnerController","description":"Owner API"}],"paths":{"/solutions":{"get":{"tags":["solution-controller"],"operationId":"getAllSolutions","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolutionDto"}}}}}}},"post":{"tags":["solution-controller"],"operationId":"addSolution","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewSolutionDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolutionDto"}}}}}}},"/owners":{"get":{"tags":["OwnerController"],"operationId":"getAllOwners","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerDto"}}}}}}},"post":{"tags":["OwnerController"],"operationId":"addOwner","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewOwnerDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerDto"}}}}}}},"/solutions/{id}":{"get":{"tags":["solution-controller"],"operationId":"getSolutionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolutionDto"}}}}}},"delete":{"tags":["solution-controller"],"operationId":"deleteSolutionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"204":{"description":"No Content"}}},"patch":{"tags":["solution-controller"],"operationId":"updateSolution","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSolutionDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolutionDto"}}}}}}},"/owners/{id}":{"get":{"tags":["OwnerController"],"operationId":"getOwnerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerDto"}}}}}},"delete":{"tags":["OwnerController"],"operationId":"deleteOwnerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"204":{"description":"No Content"}}},"patch":{"tags":["OwnerController"],"operationId":"updateOwner","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOwnerDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerDto"}}}}}}},"/tracktime":{"get":{"tags":["track-time-controller"],"operationId":"getAll","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrackTimeDto"}}}}}}}},"/tracktime/{name}":{"get":{"tags":["track-time-controller"],"operationId":"getAllByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrackTimeDto"}}}}}}}},"/tracktime/stats/{name}":{"get":{"tags":["track-time-controller"],"operationId":"getStatByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrackTimeStatDto"}}}}}}},"/tracktime/average/{name}":{"get":{"tags":["track-time-controller"],"operationId":"getAverageTrackTimeByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Error"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrackTimeAverageDto"}}}}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"errors":{"type":"array","description":"Список стектрейсов или описания ошибок","items":{"type":"string","description":"Список стектрейсов или описания ошибок"}},"message":{"type":"string","description":"сообщение об ошибке","example":"validation email size error"},"reason":{"type":"string","description":"причина возникновения","example":"Request parameters validation error"},"httpStatus":{"type":"string","description":"HTTP статус","example":"409 CONFLICT","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 EARLY_HINTS","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"timeStamp":{"type":"string","description":"Время события","format":"date-time"}},"description":"Данное о возникшей ошибке"},"NewSolutionDto":{"required":["name"],"type":"object","properties":{"name":{"maxLength":30,"minLength":2,"pattern":"^(?=.*[a-zA-Zа-яёЁА-Я\\d_\\S]).+$","type":"string","description":"Наименование продукта","example":"Синий"},"version":{"type":"string","description":"Версия продукта","example":"0.1.0"},"ownerId":{"type":"integer","description":"id владельца продукта","format":"int64","example":2}},"description":"DTO создания нового продукта"},"OwnerDto":{"type":"object","properties":{"id":{"type":"integer","description":"id владельца продукта","format":"int64","example":1},"name":{"type":"string","description":"имя владельца продукта","example":"Джесси Пинкман"},"email":{"type":"string","description":"электронная почта владельца продукта","example":"albuqerke@ya.ru"}},"description":"DTO владельца продукта","example":"   id: 1,\n   name: xxcxc,\n   email: sdsd@sdsd.fd\n"},"SolutionDto":{"type":"object","properties":{"id":{"type":"integer","description":"id продукта","format":"int64","example":1},"name":{"type":"string","description":"имя продукта","example":"Мёд"},"version":{"type":"string","description":"версия продукта","example":"0.1.1"},"ownerDto":{"$ref":"#/components/schemas/OwnerDto"}},"description":"DTO продукта"},"NewOwnerDto":{"required":["email"],"type":"object","properties":{"name":{"maxLength":30,"minLength":2,"pattern":"^(?=.*[a-zA-Zа-яёЁА-Я\\d_\\S]).+$","type":"string","description":"Имя нового владельца продукта","example":"Уолтер Уайт"},"email":{"maxLength":254,"minLength":6,"type":"string","description":"Email нового владельца продукта","example":"haisenberg@mail.ru"}},"description":"DTO создания нового владельца продукта"},"UpdateSolutionDto":{"type":"object","properties":{"name":{"maxLength":30,"minLength":2,"pattern":"^(?=.*[a-zA-Zа-яёЁА-Я\\d_\\S]).+$","type":"string","description":"Наименование продукта","example":"Синий"},"version":{"type":"string","description":"Версия продукта"},"ownerId":{"type":"integer","format":"int64"}},"description":"DTO для обновления информации о продукте"},"UpdateOwnerDto":{"type":"object","properties":{"name":{"type":"string","description":"имя владельца продукта","example":"Джесси Пинкман, beatch"},"email":{"type":"string","description":"электронная почта владельца продукта","example":"albuqerqeForever@ya.ru"}},"description":"DTO для обновления информации о владельце продукта"},"TrackTimeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"isAsync":{"type":"boolean"},"timeStamp":{"type":"string","format":"date-time"},"executionDuration":{"type":"integer","format":"int64"}}},"TrackTimeStatDto":{"type":"object","properties":{"name":{"type":"string"},"numCalls":{"type":"integer","format":"int64"},"avgDuration":{"type":"number","format":"double"},"minDuration":{"type":"integer","format":"int64"},"maxDuration":{"type":"integer","format":"int64"}}},"TrackTimeAverageDto":{"type":"object","properties":{"name":{"type":"string"},"averageDuration":{"type":"number","format":"double"}}}}}}